POM LOD
This implements a POM LOD (fades over distance) that is based on setting
a LOD bias int pomLODThreshold = 3;, which specifies the mip level at
which the fade happens.  It also has to be told the texture dimensions,
so it can do some calculations.  This is better then all the aliasing
and noise from POM in the distance, but IMO the fade feels very abrupt
because it's happening over a pretty short band.